collectors/crypto/market-collector/requirements.txt
websocket-client==1.7.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv